requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
pydantic==2.8.2
neo4j==5.22.0
qdrant-client==1.11.1
langchain==0.2.12
langchain-openai==0.1.21
python-dotenv==1.0.1
websockets==13.0
tqdm==4.66.4
requests==2.32.3
spacy==3.7.5
sentence-transformers==3.0.1
langchain-community==0.2.11
PyPDF2==3.0.1
python-docx==1.1.2
pandas==2.2.2
faiss-cpu==1.8.0
numpy==1.26.4
pillow==10.4.0
torch==2.4.0
torchvision==0.19.0
torchaudio==2.4.0
pytesseract==0.3.13
httpx==0.27.0
pytest==8.3.2
ezdxf==1.4.2
ifcopenshell==0.8.3.post2
scipy==1.14.0
sentry-sdk==2.13.0
prometheus-client==0.20.0
prometheus-fastapi-instrumentator==7.0.0
python-telegram-bot==21.6
google-api-python-client==2.140.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
pyjwt==2.9.0
slowapi==0.1.9
beautifulsoup4==4.13.5
apscheduler==3.11.0
celery==5.3.4
redis==4.5.4
psutil==5.9.8
tiktoken==0.11.0langchain==0.1.0
sentence-transformers==2.2.2
peft==0.6.0
pydantic==2.0
python-dotenv==1.0
